// apps/builder/src/api/ws/ZWEB_PROTO.ts
export interface MovingMessageBin {
    /**
     * @generated from protobuf field: tutorial.Signal signal = 1;
     */
    signal: Signal; // see package ws const define
    /**
     * @generated from protobuf field: tutorial.Target target = 2;
     */
    target: Target;
    /**
     * @generated from protobuf field: string clientID = 3;
     */
    clientID: string; // message sender client ID
    /**
     * @generated from protobuf field: bool needBroadcast = 4;
     */
    needBroadcast: boolean;
    /**
     * @generated from protobuf field: string userID = 5;
     */
    userID: string; // message sender user ID
    /**
     * @generated from protobuf field: string nickname = 6;
     */
    nickname: string; // message sender nickname
    /**
     * @generated from protobuf field: int32 status = 7;
     */
    status: number; // set instance to status, 0 for empty status
    /**
     * @generated from protobuf field: string parentDisplayName = 8;
     */
    parentDisplayName: string; // parent component displayname
    /**
     * @generated from protobuf field: string displayNames = 9;
     */
    displayNames: string; // message affected component display names, separate with comma ","
    /**
     * @generated from protobuf field: uint32 cursorXInteger = 10;
     */
    cursorXInteger: number; //  cursor's position with dot number
    /**
     * @generated from protobuf field: uint32 cursorYInteger = 11;
     */
    cursorYInteger: number; //  cursor's position with dot number
    /**
     * @generated from protobuf field: float cursorXMod = 12;
     */
    cursorXMod: number; //  cursor's position with dot number
    /**
     * @generated from protobuf field: float cursorYMod = 13;
     */
    cursorYMod: number; //  cursor's position with dot number
    /**
     * @generated from protobuf field: uint32 widgetX = 14;
     */
    widgetX: number; //  widget'position with dot number
    /**
     * @generated from protobuf field: uint32 widgetY = 15;
     */
    widgetY: number; //  widget'position with dot number
    /**
     * @generated from protobuf field: uint32 widgetW = 16;
     */
    widgetW: number; //  widget'shape with dot number
    /**
     * @generated from protobuf field: uint32 widgetH = 17;
     */
    widgetH: number; //  widget'shape with dot number
}
/**
 * @generated from protobuf enum tutorial.Signal
 */
export enum Signal {
    /**
     * @generated from protobuf enum value: SIGNAL_PING = 0;
     */
    PING = 0,
    /**
     * @generated from protobuf enum value: SIGNAL_ENTER = 1;
     */
    ENTER = 1,
    /**
     * @generated from protobuf enum value: SIGNAL_LEAVE = 2;
     */
    LEAVE = 2,
    /**
     * @generated from protobuf enum value: SIGNAL_CREATE_STATE = 3;
     */
    CREATE_STATE = 3,
    /**
     * @generated from protobuf enum value: SIGNAL_DELETE_STATE = 4;
     */
    DELETE_STATE = 4,
    /**
     * @generated from protobuf enum value: SIGNAL_UPDATE_STATE = 5;
     */
    UPDATE_STATE = 5,
    /**
     * @generated from protobuf enum value: SIGNAL_MOVE_STATE = 6;
     */
    MOVE_STATE = 6,
    /**
     * @generated from protobuf enum value: SIGNAL_CREATE_OR_UPDATE_STATE = 7;
     */
    CREATE_OR_UPDATE_STATE = 7,
    /**
     * @generated from protobuf enum value: SIGNAL_BROADCAST_ONLY = 8;
     */
    BROADCAST_ONLY = 8,
    /**
     * @generated from protobuf enum value: SIGNAL_PUT_STATE = 9;
     */
    PUT_STATE = 9,
    /**
     * @generated from protobuf enum value: SIGNAL_GLOBAL_BROADCAST_ONLY = 10;
     */
    GLOBAL_BROADCAST_ONLY = 10,
    /**
     * @generated from protobuf enum value: SIGNAL_COOPERATE_ATTACH = 11;
     */
    COOPERATE_ATTACH = 11,
    /**
     * @generated from protobuf enum value: SIGNAL_COOPERATE_DISATTACH = 12;
     */
    COOPERATE_DISATTACH = 12,
    /**
     * @generated from protobuf enum value: SIGNAL_MOVE_CURSOR = 13;
     */
    MOVE_CURSOR = 13,
    /**
     * @generated from protobuf enum value: SIGNAL_FORCE_REFRESH = 14;
     */
    FORCE_REFRESH = 14,
    /**
     * @generated from protobuf enum value: SIGNAL_TAKE_APP_SNAPSHOT = 15;
     */
    TAKE_APP_SNAPSHOT = 15,
    /**
     * @generated from protobuf enum value: SIGNAL_RECOVER_APP_SNAPSHOT = 16;
     */
    RECOVER_APP_SNAPSHOT = 16
}
/**
 * @generated from protobuf enum tutorial.Target
 */
export enum Target {
    /**
     * placeholder for nothing
     *
     * @generated from protobuf enum value: TARGET_NOTHING = 0;
     */
    NOTHING = 0,
    /**
     * ComponentsState
     *
     * @generated from protobuf enum value: TARGET_COMPONENTS = 1;
     */
    COMPONENTS = 1,
    /**
     * DependenciesState
     *
     * @generated from protobuf enum value: TARGET_DEPENDENCIES = 2;
     */
    DEPENDENCIES = 2,
    /**
     * DragShadowState
     *
     * @generated from protobuf enum value: TARGET_DRAG_SHADOW = 3;
     */
    DRAG_SHADOW = 3,
    /**
     * DottedLineSquareState
     *
     * @generated from protobuf enum value: TARGET_DOTTED_LINE_SQUARE = 4;
     */
    DOTTED_LINE_SQUARE = 4,
    /**
     * DisplayNameState
     *
     * @generated from protobuf enum value: TARGET_DISPLAY_NAME = 5;
     */
    DISPLAY_NAME = 5,
    /**
     * only for broadcast
     *
     * @generated from protobuf enum value: TARGET_APPS = 6;
     */
    APPS = 6,
    /**
     * only for broadcast
     *
     * @generated from protobuf enum value: TARGET_RESOURCE = 7;
     */
    RESOURCE = 7,
    /**
     * only for broadcast
     *
     * @generated from protobuf enum value: TARGET_ACTION = 8;
     */
    ACTION = 8,
    /**
     * only for broadcast
     *
     * @generated from protobuf enum value: TARGET_CURSOR = 9;
     */
    CURSOR = 9,
    /**
     * only for broadcast
     *
     * @generated from protobuf enum value: TARGET_WINDOW = 10;
     */
    WINDOW = 10
}
// ...
export const MovingMessageBin = new MovingMessageBin$Type();